refactor raicontours
